fix: avoid reusing instances for classes reloaded in development
This only affects development, when using threaded servers like Puma
AND with concurrency enabled.

The bug manifests as serializers "not being updated", and the cause is
that  even though the class itself is being reloaded correctly, the
`instance` method will a previously cached instance of a previous
version of the class.

By using `object_id`, if the class was reloaded, a new instance will be
created.